Move rpm::TransactionCallbacks class to separate header file
