+++ dnfdaemon-server/transaction.cpp
@@ -20,6 +20,7 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 #include "transaction.hpp"
 
 #include <libdnf/common/exception.hpp>
+#include <libdnf/rpm/transaction_callbacks.hpp>
 
 #include <type_traits>
