+++ include/libdnf/rpm/transaction.hpp
@@ -26,6 +26,7 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 #include "libdnf/base/base_weak.hpp"
 #include "libdnf/base/transaction_package.hpp"
 #include "libdnf/common/exception.hpp"
+#include "libdnf/rpm/transaction_callbacks.hpp"
 
 #include <memory>
 
@@ -63,55 +64,6 @@ private:
     explicit RpmHeader(void * hdr);
     void * header;
 };
-
-
-// suppress "unused-parameter" warnings because TransactionCallbacks is a virtual class
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wunused-parameter"
-
-/// Base class for Transaction callbacks
-/// User implements Transaction callbacks by inheriting this class and overriding its methods.
-class TransactionCallbacks {
-public:
-    /// Scriptlet type
-    // TODO(jrohel): Are all scriptlets types present and correct?
-    enum class ScriptType {
-        UNKNOWN,
-        PRE_INSTALL,            // "%pre"
-        POST_INSTALL,           // "%post"
-        PRE_UNINSTALL,          // "%preun"
-        POST_UNINSTALL,         // "%postun"
-        PRE_TRANSACTION,        // "%pretrans"
-        POST_TRANSACTION,       // "%posttrans"
-        TRIGGER_PRE_INSTALL,    // "%triggerprein"
-        TRIGGER_INSTALL,        // "%triggerin"
-        TRIGGER_UNINSTALL,      // "%triggerun"
-        TRIGGER_POST_UNINSTALL  // "%triggerpostun"
-    };
-
-    virtual ~TransactionCallbacks() = default;
-
-    virtual void install_progress(const TransactionItem & item, uint64_t amount, uint64_t total) {}
-    virtual void install_start(const TransactionItem & item, uint64_t total) {}
-    virtual void install_stop(const TransactionItem & item, uint64_t amount, uint64_t total) {}
-    virtual void transaction_progress(uint64_t amount, uint64_t total) {}
-    virtual void transaction_start(uint64_t total) {}
-    virtual void transaction_stop(uint64_t total) {}
-    virtual void uninstall_progress(const TransactionItem & item, uint64_t amount, uint64_t total) {}
-    virtual void uninstall_start(const TransactionItem & item, uint64_t total) {}
-    virtual void uninstall_stop(const TransactionItem & item, uint64_t amount, uint64_t total) {}
-    virtual void unpack_error(const TransactionItem & item) {}
-    virtual void cpio_error(const TransactionItem & item) {}
-    virtual void script_error(const TransactionItem * item, Nevra nevra, ScriptType type, uint64_t return_code) {}
-    virtual void script_start(const TransactionItem * item, Nevra nevra, ScriptType type) {}
-    virtual void script_stop(const TransactionItem * item, Nevra nevra, ScriptType type, uint64_t return_code) {}
-    virtual void elem_progress(const TransactionItem & item, uint64_t amount, uint64_t total) {}
-    virtual void verify_progress(uint64_t amount, uint64_t total) {}
-    virtual void verify_start(uint64_t total) {}
-    virtual void verify_stop(uint64_t total) {}
-};
-
-#pragma GCC diagnostic pop
 
 
 class RpmProblem {
